moved FileIO to sciencebeam_gym.utils.tf
