+++ sciencebeam_gym/tools/colorize_image.py
@@ -7,7 +7,7 @@ import re
 
 from six.moves.configparser import ConfigParser
 
-from sciencebeam_gym.trainer.util import FileIO
+from sciencebeam_gym.utils.tf import FileIO
 
 def get_args_parser():
   parser = argparse.ArgumentParser()
